requirements.txt
Flask==2.2.5
face_recognition==1.3.0
opencv-python==4.5.5.64
Pillow==8.4.0
numpy==1.21.4
